Add device controller to the default configuration
When the cgroup-bin package is installed, it mounts two cgroups.  Then
whenever libvirt tries to fire up a VM, it sees cgroups are mounted, and
wants to lock the VM into a devices cgroup to stop it accessing devices.
However the default config does not mount the devices cgroup, so libvirt
refuses to start up the VM.

LP: #607849
Thanks: Serge Hallyn