Remove initscripts cgconfig and cgred
This means that available cgroups are not mounted during boot and no process
classification will occur.  They are now located in the examples/ directory of
cgroup-bin's documentation.

I realize this may be an unpopular decision, but the initscripts are currently
plagued with problems, including race conditions, collisions with lxc and
libvirt, and crashes on non-systemd configurations.  It's possible they will be
restored in the future as time permits, but this at least gives me a stable
point from which to move forward.